Replaced double for floats

+++ src/Constants.hpp
@@ -6,7 +6,7 @@ const int WINDOW_WIDTH = 800;
 const int WINDOW_HEIGHT = 600;
 
 // World constants
-const double G = 6.67E-11;
-const double TIME_DELTA = 0.01;
+const float G = 6.67E-11;
+const float TIME_DELTA = 0.01;
 
 #endif
